api/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
PyMySQL==1.1.0
python-dotenv==1.0.0
marshmallow==3.20.1
bcrypt==4.0.1
Flask-Session==0.5.0
pytest==7.4.2
pytest-flask==1.2.0
Werkzeug==2.3.7
SQLAlchemy==2.0.21
click==8.1.7
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
attrs==23.1.0
exceptiongroup==1.1.3
iniconfig==2.0.0
packaging==23.1
pluggy==1.3.0
tomli==2.0.1
cryptography==41.0.4
cffi==1.15.1
pycparser==2.21